game/backend/requirements.txt
channels==4.2.0
Django==5.1.5
djangorestframework==3.15.2
channels_redis==4.2.1
